+++ eZ/Publish/Core/Repository/SectionService.php
@@ -24,8 +24,9 @@ use eZ\Publish\API\Repository\Values\Content\SectionCreateStruct,
     eZ\Publish\Core\Base\Exceptions\InvalidArgumentValue,
     eZ\Publish\Core\Base\Exceptions\InvalidArgumentException,
     eZ\Publish\Core\Base\Exceptions\BadStateException,
-    eZ\Publish\API\Repository\Exceptions\NotFoundException,
-    eZ\Publish\API\Repository\Exceptions\UnauthorizedException;
+    eZ\Publish\Core\Base\Exceptions\UnauthorizedException,
+
+    eZ\Publish\API\Repository\Exceptions\NotFoundException as APINotFoundException;
 
 /**
  * Section service, used for section operations
@@ -81,21 +82,34 @@ class SectionService implements SectionServiceInterface
         if ( !is_string( $sectionCreateStruct->identifier ) || empty( $sectionCreateStruct->identifier ) )
             throw new InvalidArgumentValue( "identifier", $sectionCreateStruct->identifier, "SectionCreateStruct" );
 
+        if ( $this->repository->hasAccess( 'section', 'edit' ) !== true )
+            throw new UnauthorizedException( 'section', 'edit' );
+
         try
         {
             $existingSection = $this->loadSectionByIdentifier( $sectionCreateStruct->identifier );
             if ( $existingSection !== null )
                 throw new InvalidArgumentException( "sectionCreateStruct", "section with specified identifier already exists" );
         }
-        catch ( NotFoundException $e )
+        catch ( APINotFoundException $e )
         {
             // Do nothing
         }
 
-        $spiSection = $this->persistenceHandler->sectionHandler()->create(
-            $sectionCreateStruct->name,
-            $sectionCreateStruct->identifier
-        );
+        $this->repository->beginTransaction();
+        try
+        {
+            $spiSection = $this->persistenceHandler->sectionHandler()->create(
+                $sectionCreateStruct->name,
+                $sectionCreateStruct->identifier
+            );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
 
         return $this->buildDomainSectionObject( $spiSection );
     }
@@ -122,6 +136,9 @@ class SectionService implements SectionServiceInterface
         if ( $sectionUpdateStruct->identifier !== null && !is_string( $sectionUpdateStruct->identifier ) )
             throw new InvalidArgumentValue( "identifier", $section->identifier, "Section" );
 
+        if ( $this->repository->canUser( 'section', 'edit', $section ) !== true )
+            throw new UnauthorizedException( 'section', 'edit' );
+
         if ( $sectionUpdateStruct->identifier !== null )
         {
             try
@@ -130,7 +147,7 @@ class SectionService implements SectionServiceInterface
                 if ( $existingSection !== null )
                     throw new InvalidArgumentException( "sectionUpdateStruct", "section with specified identifier already exists" );
             }
-            catch ( NotFoundException $e )
+            catch ( APINotFoundException $e )
             {
                 // Do nothing
             }
@@ -138,11 +155,21 @@ class SectionService implements SectionServiceInterface
 
         $loadedSection = $this->loadSection( $section->id );
 
-        $spiSection = $this->persistenceHandler->sectionHandler()->update(
-            $loadedSection->id,
-            $sectionUpdateStruct->name ?: $loadedSection->name,
-            $sectionUpdateStruct->identifier ?: $loadedSection->identifier
-        );
+        $this->repository->beginTransaction();
+        try
+        {
+            $spiSection = $this->persistenceHandler->sectionHandler()->update(
+                $loadedSection->id,
+                $sectionUpdateStruct->name ?: $loadedSection->name,
+                $sectionUpdateStruct->identifier ?: $loadedSection->identifier
+            );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
 
         return $this->buildDomainSectionObject( $spiSection );
     }
@@ -162,6 +189,9 @@ class SectionService implements SectionServiceInterface
         if ( !is_numeric( $sectionId ) )
             throw new InvalidArgumentValue( "sectionId", $sectionId );
 
+        if ( $this->repository->hasAccess( 'section', 'view' ) !== true )
+            throw new UnauthorizedException( 'section', 'view' );
+
         $spiSection = $this->persistenceHandler->sectionHandler()->load( $sectionId );
         return $this->buildDomainSectionObject( $spiSection );
     }
@@ -175,6 +205,9 @@ class SectionService implements SectionServiceInterface
      */
     public function loadSections()
     {
+        if ( $this->repository->hasAccess( 'section', 'view' ) !== true )
+            throw new UnauthorizedException( 'section', 'view' );
+
         $spiSections = $this->persistenceHandler->sectionHandler()->loadAll();
 
         $sections = array();
@@ -200,6 +233,9 @@ class SectionService implements SectionServiceInterface
     {
         if ( !is_string( $sectionIdentifier ) || empty( $sectionIdentifier ) )
             throw new InvalidArgumentValue( "sectionIdentifier", $sectionIdentifier );
+
+        if ( $this->repository->hasAccess( 'section', 'view' ) !== true )
+            throw new UnauthorizedException( 'section', 'view' );
 
         $spiSection = $this->persistenceHandler->sectionHandler()->loadByIdentifier( $sectionIdentifier );
         return $this->buildDomainSectionObject( $spiSection );
@@ -240,10 +276,23 @@ class SectionService implements SectionServiceInterface
         $loadedContentInfo = $this->repository->getContentService()->loadContentInfo( $contentInfo->id );
         $loadedSection = $this->loadSection( $section->id );
 
-        $this->persistenceHandler->sectionHandler()->assign(
-            $loadedSection->id,
-            $loadedContentInfo->id
-        );
+        if ( $this->repository->canUser( 'section', 'assign', $loadedContentInfo, $loadedSection ) !== true )
+            throw new UnauthorizedException( 'section', 'assign', $loadedSection->id );
+
+        $this->repository->beginTransaction();
+        try
+        {
+            $this->persistenceHandler->sectionHandler()->assign(
+                $loadedSection->id,
+                $loadedContentInfo->id
+            );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
     }
 
     /**
@@ -263,10 +312,23 @@ class SectionService implements SectionServiceInterface
 
         $loadedSection = $this->loadSection( $section->id );
 
+        if ( $this->repository->canUser( 'section', 'edit', $loadedSection ) !== true )
+            throw new UnauthorizedException( 'section', 'edit', $loadedSection->id );
+
         if ( $this->countAssignedContents( $loadedSection ) > 0 )
             throw new BadStateException( "section", 'section is still assigned to content' );
 
-        $this->persistenceHandler->sectionHandler()->delete( $loadedSection->id );
+        $this->repository->beginTransaction();
+        try
+        {
+            $this->persistenceHandler->sectionHandler()->delete( $loadedSection->id );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
     }
 
     /**
