+++ eZ/Publish/Core/Repository/TrashService.php
@@ -23,10 +23,13 @@ use eZ\Publish\API\Repository\TrashService as TrashServiceInterface,
     eZ\Publish\SPI\Persistence\Content\Location\Trashed,
 
     eZ\Publish\Core\Base\Exceptions\InvalidArgumentValue,
+    eZ\Publish\Core\Base\Exceptions\UnauthorizedException,
 
     eZ\Publish\API\Repository\Values\Content\SearchResult,
     eZ\Publish\API\Repository\Values\Content\Query\Criterion,
-    eZ\Publish\API\Repository\Values\Content\Query\SortClause;
+    eZ\Publish\API\Repository\Values\Content\Query\Criterion\ParentLocationId,
+    eZ\Publish\API\Repository\Values\Content\Query\SortClause,
+    DateTime;
 
 /**
  * Trash service, used for managing trashed content
@@ -101,7 +104,21 @@ class TrashService implements TrashServiceInterface
         if ( !is_numeric( $location->id ) )
             throw new InvalidArgumentValue( "id", $location->id, "Location" );
 
-        $spiTrashItem = $this->persistenceHandler->trashHandler()->trash( $location->id );
+        if ( $this->repository->canUser( 'content', 'manage_locations', $location->getContentInfo(), $location ) !== true )
+            throw new UnauthorizedException( 'content', 'manage_locations' );
+
+        $this->repository->beginTransaction();
+        try
+        {
+            $spiTrashItem = $this->persistenceHandler->trashHandler()->trash( $location->id );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
+
         return $this->buildDomainTrashItemObject( $spiTrashItem );
     }
 
@@ -128,10 +145,40 @@ class TrashService implements TrashServiceInterface
         if ( $newParentLocation !== null && !is_numeric( $newParentLocation->id ) )
             throw new InvalidArgumentValue( "parentLocationId", $newParentLocation->id, "Location" );
 
-        $newLocationId = $this->persistenceHandler->trashHandler()->recover(
-            $trashItem->id,
-            $newParentLocation ? $newParentLocation->id : $trashItem->parentLocationId
-        );
+        if ( $this->repository->hasAccess( 'content', 'restore' ) !== true )
+            throw new UnauthorizedException( 'content', 'restore' );
+
+        $this->repository->beginTransaction();
+        try
+        {
+            $newLocationId = $this->persistenceHandler->trashHandler()->recover(
+                $trashItem->id,
+                $newParentLocation ? $newParentLocation->id : $trashItem->parentLocationId
+            );
+
+            // There is a possibility for content to loose main location when one of its locations is recovered
+            // from trash, so we need to check for it and set the newly recovered location to be the main one
+            $contentService = $this->repository->getContentService();
+
+            $contentInfo = $contentService->loadContentInfo( $trashItem->contentId );
+            if ( !is_numeric( $contentInfo->mainLocationId ) )
+            {
+                $contentMetadataUpdateStruct = $contentService->newContentMetadataUpdateStruct();
+                $contentMetadataUpdateStruct->mainLocationId = $newLocationId;
+
+                $contentService->updateContentMetadata(
+                    $contentInfo,
+                    $contentMetadataUpdateStruct
+                );
+            }
+
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
 
         return $this->repository->getLocationService()->loadLocation( $newLocationId );
     }
@@ -146,8 +193,21 @@ class TrashService implements TrashServiceInterface
      */
     public function emptyTrash()
     {
-        // Persistence layer takes care of deleting content objects
-        $this->persistenceHandler->trashHandler()->emptyTrash();
+        if ( $this->repository->hasAccess( 'content', 'cleantrash' ) !== true )
+            throw new UnauthorizedException( 'content', 'cleantrash' );
+
+        $this->repository->beginTransaction();
+        try
+        {
+            // Persistence layer takes care of deleting content objects
+            $this->persistenceHandler->trashHandler()->emptyTrash();
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
     }
 
     /**
@@ -161,10 +221,23 @@ class TrashService implements TrashServiceInterface
      */
     public function deleteTrashItem( APITrashItem $trashItem )
     {
+        if ( $this->repository->hasAccess( 'content', 'cleantrash' ) !== true )
+            throw new UnauthorizedException( 'content', 'cleantrash' );
+
         if ( !is_numeric( $trashItem->id ) )
             throw new InvalidArgumentValue( "id", $trashItem->id, "TrashItem" );
 
-        $this->persistenceHandler->trashHandler()->deleteTrashItem( $trashItem->id );
+        $this->repository->beginTransaction();
+        try
+        {
+            $this->persistenceHandler->trashHandler()->deleteTrashItem( $trashItem->id );
+            $this->repository->commit();
+        }
+        catch ( \Exception $e )
+        {
+            $this->repository->rollback();
+            throw $e;
+        }
     }
 
     /**
@@ -231,6 +304,16 @@ class TrashService implements TrashServiceInterface
     {
         $contentInfo = $this->repository->getContentService()->loadContentInfo( $spiTrashItem->contentId );
 
+        $trashedChildren = array_filter(
+            $this->persistenceHandler->trashHandler()->findTrashItems(
+                new ParentLocationId( $spiTrashItem->id )
+            ),
+            function( $trashedChild ) use ( $spiTrashItem )
+            {
+                return $trashedChild->parentId === $spiTrashItem->id;
+            }
+        );
+
         return new TrashItem(
             array(
                 'contentInfo' => $contentInfo,
@@ -241,13 +324,26 @@ class TrashService implements TrashServiceInterface
                 'remoteId' => $spiTrashItem->remoteId,
                 'parentLocationId' => (int) $spiTrashItem->parentId,
                 'pathString' => $spiTrashItem->pathString,
-                'modifiedSubLocationDate' => new \DateTime( '@' . (int) $spiTrashItem->modifiedSubLocation ),
+                'modifiedSubLocationDate' => $this->getDateTime( $spiTrashItem->modifiedSubLocation ),
                 'depth' => (int) $spiTrashItem->depth,
                 'sortField' => (int) $spiTrashItem->sortField,
                 'sortOrder' => (int) $spiTrashItem->sortOrder,
-                //@todo: this has to be 0?
-                'childCount' => 0
+                'childCount' => count( $trashedChildren )
             )
         );
+    }
+
+    /**
+     *
+     *
+     * @param int|null $timestamp
+     *
+     * @return \DateTime|null
+     */
+    protected function getDateTime( $timestamp )
+    {
+        $dateTime = new DateTime();
+        $dateTime->setTimestamp( $timestamp );
+        return $dateTime;
     }
 }
